moving calculations module to different directory
